+++ native/tests/cases/knr/kr_5_8.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdlib.h>
 void swap(void *v[], int i, int j) {
     void *temp;
     temp = v[i];